requirements.txt
torch~=1.9.0
numpy~=1.19.5
scipy~=1.6.2
torchvision~=0.10.0
pillow~=8.2.0
lpips~=0.1.3
opencv-python~=4.5.3.56
pandas~=1.3.0
scikit-image~=0.18.1
seaborn~=0.11.1
argparse~=1.4.0